Creating Marketbasket
Association rule mining is the data mining process of finding the rules that may govern associations
between sets of items. The term market basket analysis refers to a specific implementation of
association rules mining that many companies use for a variety of purposes, so in a given
transaction with multiple items, it tries to find the rules that govern how or why such items are often
bought together. Association rules are rules which surpass user-specified minimum support and
minimum Confidence threshold. For example, peanut butter and jelly are often bought together
because a lot of people like to make sandwiches using these two items. Also, surprisingly, diapers
and beer are bought together because, as it turns out, that dads are often tasked to do the
shopping while the moms are left with the baby.